requirements.txt
# Core ML/Data Science
tensorflow>=2.13.0
keras>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
PyWavelets>=1.4.1

# API Framework
fastapi>=0.103.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
websockets>=11.0

# Data Collection
requests>=2.31.0
schedule>=1.2.0

# Caching & Streaming
redis>=5.0.0
kafka-python>=2.0.2

# Time-Series Database
influxdb-client>=1.36.0

# Monitoring & Metrics
prometheus-client>=0.17.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Optional: Production
gunicorn>=21.2.0  # WSGI server
celery>=5.3.0  # Task queue (optional)
flower>=2.0.0  # Celery monitoring (optional)

# Optional: Development
jupyter>=1.0.0
ipython>=8.14.0
black>=23.7.0  # Code formatter
flake8>=6.1.0  # Linter